+++ src/nextjs/templates.ts
@@ -139,6 +139,15 @@ export function getSentryConfigContents(
   ],`;
   }
 
+  let spotlightOption = '';
+  if (config === 'server') {
+    spotlightOption = `
+
+  // uncomment the line below to enable Spotlight (https://spotlightjs.com)
+  // spotlight: process.env.NODE_ENV === 'development',
+  `;
+  }
+
   // eslint-disable-next-line @typescript-eslint/restrict-template-expressions
   return `${primer}
 
@@ -151,7 +160,7 @@ Sentry.init({
   tracesSampleRate: 1,
 
   // Setting this option to true will print useful information to the console while you're setting up Sentry.
-  debug: false,${additionalOptions}
+  debug: false,${additionalOptions}${spotlightOption}
 });
 `;
 }