+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 
 ## Unreleased
 
+- feat(nextjs): Add comment to add spotlight in Sentry.init for Next.js server config (#545)
 - feat(nextjs): Pin installed Next.js SDK version to version 7 (#550)
-- feat(nextjs): Added comment to add spotlight in Sentry.init for Next.js server config (#545)
 - feat(remix): Add example page (#542)
 - feat(sveltekit): Add comment for spotlight in Sentry.init for SvelteKit server hooks config (#546)
 - ref(nextjs): Add note about `tunnelRoute` and Next.js middleware incompatibility (#544)
@@ -29,8 +29,10 @@
 
 ## 3.20.1
 
-- fix(nextjs): Replace deprecated Sentry API calls in example page templates (#520)
-- fix(sveltekit): Replace deprecated Sentry API calls in example page templates (#520)
+- fix(nextjs): Replace deprecated Sentry API calls in example page templates
+  (#520)
+- fix(sveltekit): Replace deprecated Sentry API calls in example page templates
+  (#520)
 
 ## 3.20.0